+++ src/main/java/com/desire323/authentiacation/DTO/RegisterResponse.java
@@ -1,9 +1,9 @@
 package com.desire323.authentiacation.DTO;
 public class RegisterResponse {
-    private int id;
+    private Long id;
     private String email;
 
-    public RegisterResponse(int id, String email) {
+    public RegisterResponse(Long id, String email) {
         this.id = id;
         this.email = email;
     }
@@ -11,11 +11,11 @@ public class RegisterResponse {
     public RegisterResponse() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
